# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FleetVision AI",
        "description": "FleetVision AI Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/dashboard/vehicle/{vehicle_id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Vehicle Dashboard",
                "operationId": "vehicle_dashboard_api_v1_dashboard_vehicle__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/ai/{vehicle_id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "AI Dashboard",
                "operationId": "ai_dashboard_api_v1_dashboard_ai__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/charts/speed/{vehicle_id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Speed History",
                "operationId": "speed_history_api_v1_dashboard_charts_speed__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/charts/fuel/{vehicle_id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Fuel History",
                "operationId": "fuel_history_api_v1_dashboard_charts_fuel__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/charts/temperature/{vehicle_id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Engine Temperature",
                "operationId": "engine_temperature_api_v1_dashboard_charts_temperature__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/charts/route/{vehicle_id}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "GPS Route",
                "operationId": "gps_route_api_v1_dashboard_charts_route__vehicle_id__get",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Vehicle Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    },
...
}